requirements.txt
flickrapi==2.4.0
streamlit==0.86.0
countryinfo==0.1.2
pandas==1.3.1
couchbase==3.2.0
python-dotenv==0.19.0
